% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ciEnvelope}
\alias{ciEnvelope}
\title{ciEnvelope}
\usage{
ciEnvelope(x, ylo, yhi, ...)
}
\arguments{
\item{x}{x-axis data}

\item{ylo}{y-axis lower confidence bound}

\item{yhi}{y-axis upper confidence bound}

\item{...}{optional graphical parameters}
}
\description{
plot a confidence/credible interval
}
\author{
Mike Dietze
}

