% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col.alpha}
\alias{col.alpha}
\title{add transparency to a color}
\usage{
col.alpha(col, alpha = 1)
}
\arguments{
\item{col}{color (by name, number, or hex string)}

\item{alpha}{transparancy on a 0-1 scale (1=opaque)}
}
\description{
add transparency to a color
}
\author{
Mike Dietze
}
