% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm_pois.R
\name{fit_dlm_pois}
\alias{fit_dlm_pois}
\title{fit_dlm_pois}
\usage{
fit_dlm_pois(model = NULL, data)
}
\arguments{
\item{model}{list containing the following elements
\itemize{
 \item{obs}{column name of the observed data. REQUIRED}
 \item{fixed}{formula for fixed effects. Response variable is optional but should be 'x' if included}
 \item{random}{not implemented yet; will be formula for random effects}
 \item{n.iter}{number of mcmc iterations}
}}

\item{data}{data frame containing observations and covariates}
}
\description{
Fits a Bayesian state-space dynamic linear model using JAGS
}
\author{
Mike Dietze and Andrew Tredennick
}

