% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ss}
\alias{plot_ss}
\title{plot_ss}
\usage{
plot_ss(time, fit, ...)
}
\arguments{
\item{time}{time axis values}

\item{fit}{fit_dlm output object}

\item{...}{optional graphical parameters}
}
\description{
plot time series output of a state-space model
}
\author{
Mike Dietze
}

