% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dlm_lnorm.R
\name{predict_dlm_lnorm}
\alias{predict_dlm_lnorm}
\title{predict_dlm_lnorm}
\usage{
predict_dlm_lnorm(fit, newdata = NULL, n.iter = 5000, steps = NULL,
  start.time = NULL, include = c("I", "P", "D", "E"))
}
\arguments{
\item{fit}{fit_dlm output list}

\item{newdata}{array of driver data organized with dimensions [ensemble,time,variable]}

\item{n.iter}{number of samples}

\item{steps}{number of steps}

\item{start.time}{time for start of forecast. Defaults to end of fit}

\item{include}{which source of uncertainty to include (vector of strings). Options are
\itemize{
 \item{I}{Initial Conditions}
 \item{P}{Parameters}
 \item{D}{Drivers}
 \item{E}{Process Error}
}}
}
\value{
list
\itemize{
 \item{predict}{matrix of predictions [ens,time]}
 \item{index}{data.frame of the parameter (P) and driver (D) indices used for each ensemble member}
}
}
\description{
make predictions from fit dlm for new data
}
\author{
Mike Dietze and Andrew Tredennick
}

