% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wtd.quantile}
\alias{wtd.quantile}
\title{weighted quantile}
\usage{
wtd.quantile(x, wt, q)
}
\arguments{
\item{x}{numeric vector whose quantiles are wanted}

\item{wt}{vector of weights}

\item{q}{quantile (at the moment must be scalar) 
generalization of the base quantile function to allow values to be weighted. Useful for constructing interval estimates when using a particle filter or similar algorithm}
}
\description{
weighted quantile
}
\author{
Mike Dietze
}
